backend/requirements.txt
asyncpg>=0.30.0,<1.0.0
fastapi>=0.116.0,<1.0.0
google-genai>=1.30.0,<2.0.0
numpy>=2.3.0,<3.0.0
pydantic>=2.11.0,<3.0.0
python-jose>=3.4.0,<4.0.0
python-dotenv>=1.1.1,<2.0.0
python-multipart>=0.0.20,<1.0.0
uvicorn[standard]>=0.35.0,<1.0.0
